Removing unused class
